# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Signara API
  version: 1.0.0
  description: |
    Secure Every Signature. — REST API for the Signara digital document signing and
    agreement management platform.

    - Base path: `/api/v1`
    - Interactive docs (Swagger UI): `/api/v1/docs`
    - Live YAML: `/api/v1/openapi.yaml`

    ## Authentication
    Two mechanisms are supported:

    | Scheme | Where | Usage |
    | --- | --- | --- |
    | Bearer JWT (`access-token`) | `Authorization: Bearer <token>` | Interactive sessions issued by the OIDC flow |
    | API key (`api-key`) | `X-API-Key: sgn_...` | Machine-to-machine clients |

    ## Tenancy
    Tenant-scoped endpoints require an active organization context. Endpoints decorated
    with `x-tenant-required: true` return `403` when the caller has no membership.

    ## Idempotency
    Mutating endpoints accept `X-Idempotency-Key` — repeat requests with the same key
    return the stored response instead of re-executing.
  termsOfService: https://signara.innotel.us/terms
  contact:
    name: Signara
    url: https://signara.innotel.us
    email: support@signara.innotel.us
  license:
    name: AGPL-3.0-or-later
    identifier: AGPL-3.0-or-later
servers:
  - url: https://api.signara.innotel.us/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local development
...
security:
  - access-token: []
paths:
...
  /signatures/requests:
    get:
      tags: [signatures]
      x-tenant-required: true
      summary: List signing requests
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/DocumentStatus' } }
        - { name: limit, in: query, schema: { type: integer, default: 25, maximum: 100 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: Paginated requests
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/SigningRequestSummary' }
    post:
      tags: [signatures]
      x-tenant-required: true
      summary: Create a signing request (sequential or parallel)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SigningRequestCreate' }
      responses:
        '201':
          description: Request created with signer sessions
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SigningRequestDetail' }
        '409': { $ref: '#/components/responses/Conflict' }
  /signatures/requests/{id}:
    get:
      tags: [signatures]
      x-tenant-required: true
      summary: Get a signing request with events and signatures
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Request detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SigningRequestDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
  /signatures/requests/{id}/cancel:
    post:
      tags: [signatures]
      x-tenant-required: true
      summary: Cancel a signing request
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        '200':
          description: Request cancelled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Success' }
  /signatures/requests/{id}/remind:
    post:
      tags: [signatures]
      x-tenant-required: true
      summary: Send reminders to pending signers (max once per 24h per signer)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                signerId: { type: string, format: uuid, nullable: true }
      responses:
        '200':
          description: Reminders enqueued
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  enqueued: { type: integer }
  /signatures/requests/{id}/evidence:
    get:
      tags: [signatures]
      x-tenant-required: true
      summary: Generate the signature evidence report
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Evidence report
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EvidenceReport' }
  /signatures/public/{token}:
    get:
      tags: [signatures]
      security: []
      summary: Public signing session (unguessable per-signer token)
      description: No authentication — the token is the credential.
      parameters:
        - { name: token, in: path, required: true, schema: { type: string, example: sgn_xxxx } }
      responses:
        '200':
          description: Signing session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SigningSession' }
        '404': { $ref: '#/components/responses/NotFound' }
  /signatures/public/{token}/sign:
    post:
      tags: [signatures]
      security: []
      summary: Record a signature from the public session
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignInput' }
      responses:
        '200':
          description: Signature recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  signatureId: { type: string, format: uuid }
        '409': { $ref: '#/components/responses/Conflict' }
  /signatures/public/{token}/decline:
    post:
      tags: [signatures]
      security: []
      summary: Decline to sign
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        '200':
          description: Decline recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Success' }
  /signatures/public/{token}/events:
    get:
      tags: [signatures]
      security: []
      summary: Event history visible to the signer
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Signer-visible events
          content:
            application/json:
              schema:
                type: object
                properties:
                  requestId: { type: string, format: uuid }
                  signerId: { type: string, format: uuid }
                  events:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        type: { $ref: '#/components/schemas/SignatureEventType' }
                        createdAt: { type: string, format: date-time }
                        metadata: { type: object }

  # --------------------------------------------------------------- audit ---
...
components:
  securitySchemes:
    access-token:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by the Signara API after the Authentik OIDC flow, or the raw Authentik access token.
    api-key:
      type: apiKey
      in: header
      name: X-API-Key
      description: "API key created via POST /api-keys, e.g. `X-API-Key: sgn_...`"
  responses:
...
    NotFound:
      description: Resource does not exist or is outside the tenant
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: State conflict (e.g. duplicate email, already signed)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
    Success:
      type: object
      properties:
        success: { type: boolean, example: true }
    ErrorEnvelope:
      type: object
      properties:
        statusCode: { type: integer }
        error: { type: string }
        message:
          type: array
          items: { type: string }
        path: { type: string }
        timestamp: { type: string, format: date-time }
        requestId: { type: string }

    # ------------------------------------------------------------- auth ---
...
    DocumentStatus:
      type: string
      enum: [DRAFT, AWAITING_SIGNATURE, IN_PROGRESS, COMPLETED, VOIDED, CANCELLED, EXPIRED]
...
    SigningMode:
      type: string
      enum: [SEQUENTIAL, PARALLEL]
    SignerRole:
      type: string
      enum: [SIGNER, APPROVER, CC]
    SignerStatus:
      type: string
      enum: [PENDING, INVITED, VIEWED, SIGNED, DECLINED, RESCINDED, EXPIRED]
    SignatureType:
      type: string
      enum: [TYPED, DRAWN, UPLOADED, CERTIFICATE]
    SignatureEventType:
      type: string
      enum: [CREATED, INVITED, VIEWED, SIGNED, DECLINED, REMINDED, EXPIRED, CANCELLED, VOIDED, APPROVED, REJECTED, DOWNLOADED, EMAIL_FAILED]
    SigningRequestCreate:
      type: object
      required: [documentId, signers]
      properties:
        documentId: { type: string, format: uuid }
        title: { type: string }
        message: { type: string }
        deadline: { type: string, format: date-time }
        mode: { $ref: '#/components/schemas/SigningMode' }
        sendInvites: { type: boolean, default: true }
        signers:
          type: array
          minItems: 1
          maxItems: 50
          items:
            type: object
            required: [email]
            properties:
              email: { type: string, format: email }
              name: { type: string }
              role: { $ref: '#/components/schemas/SignerRole' }
              orderIndex: { type: integer, minimum: 0 }
        workflowRules:
          type: array
          items:
            type: object
            properties:
              orderIndex: { type: integer, minimum: 0 }
              condition: { type: object }
              action: { type: string, enum: [APPROVE, ROUTE, REQUIRE, NOTIFY] }
              targetSignerId: { type: string, format: uuid }
    SigningRequestSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        organizationId: { type: string, format: uuid }
        documentId: { type: string, format: uuid }
        title: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/DocumentStatus' }
        mode: { $ref: '#/components/schemas/SigningMode' }
        deadline: { type: string, format: date-time, nullable: true }
        completedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        document:
          type: object
          properties:
            id: { type: string, format: uuid }
            title: { type: string }
            fileName: { type: string }
    SigningRequestDetail:
      allOf:
        - $ref: '#/components/schemas/SigningRequestSummary'
        - type: object
          properties:
            message: { type: string, nullable: true }
            signers:
              type: array
              items:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  email: { type: string, format: email }
                  name: { type: string, nullable: true }
                  role: { $ref: '#/components/schemas/SignerRole' }
                  orderIndex: { type: integer }
                  status: { $ref: '#/components/schemas/SignerStatus' }
                  viewedAt: { type: string, format: date-time, nullable: true }
                  signedAt: { type: string, format: date-time, nullable: true }
            signatures:
              type: array
              items: { $ref: '#/components/schemas/Signature' }
            events:
              type: array
              items:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  type: { $ref: '#/components/schemas/SignatureEventType' }
                  signerId: { type: string, format: uuid, nullable: true }
                  metadata: { type: object, nullable: true }
                  ipAddress: { type: string, nullable: true }
                  createdAt: { type: string, format: date-time }
    Signature:
      type: object
      properties:
        id: { type: string, format: uuid }
        requestId: { type: string, format: uuid }
        documentId: { type: string, format: uuid }
        signerId: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/SignatureType' }
        certificateId: { type: string, format: uuid, nullable: true }
        certificateSerial: { type: string, nullable: true }
        signedHash: { type: string, nullable: true }
        signatureValue: { type: string, nullable: true, description: Detached cryptographic signature (base64) }
        signatureFormat: { type: string, nullable: true, example: RSA-SHA256 }
        cryptoAlgorithm: { type: string, nullable: true }
        identityAssurance: { $ref: '#/components/schemas/IdentityAssurance' }
        ipAddress: { type: string, nullable: true }
        signedAt: { type: string, format: date-time }
        status: { type: string, enum: [VALID, INVALID, REVOKED] }
    IdentityAssurance:
      type: object
      properties:
        score:
          type: integer
          minimum: 0
          maximum: 100
          description: Transparent factor-weighted score (see docs/Security.md)
        level:
          type: string
          enum: [NONE, BASIC, ADVANCED, HIGH]
        factors:
          type: object
          properties:
            certificateValid: { type: boolean }
            certIdentityMatch: { type: boolean }
            accountVerified: { type: boolean }
            mfaEnabled: { type: boolean }
            revocationChecked: { type: boolean }
            providerValidation: { type: string, nullable: true, enum: [DV, OV, EV] }
...
    SignInput:
      type: object
      properties:
        type: { $ref: '#/components/schemas/SignatureType' }
        certificateSerial: { type: string }
        signedHash: { type: string, description: SHA-256 hex binding document + signer + request }
        cryptoAlgorithm: { type: string, default: SHA-256 }
        signatureData: { type: string }
        certificateId: { type: string, format: uuid, description: Certificate to bind (required for CERTIFICATE signatures) }
        signatureValue: { type: string, description: Base64 cryptographic signature produced by the key holder }
        signatureFormat: { type: string, example: RSA-SHA256 }
    SigningSession:
      type: object
      properties:
        requestId: { type: string, format: uuid }
        title: { type: string }
        document:
          type: object
          properties:
            id: { type: string, format: uuid }
            fileName: { type: string }
            downloadUrl: { type: string, format: uri }
        signer:
          type: object
          properties:
            id: { type: string, format: uuid }
            email: { type: string, format: email }
            name: { type: string, nullable: true }
            role: { $ref: '#/components/schemas/SignerRole' }
            status: { $ref: '#/components/schemas/SignerStatus' }
            orderIndex: { type: integer }
        message: { type: string, nullable: true }
        deadline: { type: string, format: date-time, nullable: true }
        mode: { $ref: '#/components/schemas/SigningMode' }
        allowsSigning: { type: boolean }
        authMethod: { type: string, enum: [email, oidc, certificate] }
        requestedFields:
          type: array
          description: Field definitions when signing from a template
    EvidenceReport:
      type: object
      properties:
        reportId: { type: string, format: uuid }
        generatedAt: { type: string, format: date-time }
        requestId: { type: string, format: uuid }
        document:
          type: object
          properties:
            id: { type: string, format: uuid }
            title: { type: string }
            fileName: { type: string }
            checksumSha256: { type: string, nullable: true }
        mode: { $ref: '#/components/schemas/SigningMode' }
        status: { $ref: '#/components/schemas/DocumentStatus' }
        envelope:
          type: object
          properties:
            signers:
              type: array
              items:
                type: object
                properties:
                  email: { type: string, format: email }
                  name: { type: string, nullable: true }
                  role: { $ref: '#/components/schemas/SignerRole' }
                  status: { $ref: '#/components/schemas/SignerStatus' }
                  signedAt: { type: string, format: date-time, nullable: true }
                  authMethod: { type: string, nullable: true }
                  signatures:
                    type: array
                    items: { $ref: '#/components/schemas/Signature' }
        auditTrail:
          type: array
          items:
            type: object
            properties:
              type: { $ref: '#/components/schemas/SignatureEventType' }
              at: { type: string, format: date-time }
              signerId: { type: string, format: uuid, nullable: true }
              metadata: { type: object, nullable: true }
              ipAddress: { type: string, nullable: true }
              userAgent: { type: string, nullable: true }
        statement: { type: string }

    # --------------------------------------------------------------- audit -
...